NEW ADDITION! Quotes from my local API
This is a feature where the users will be welcome to a random quote out
of ten quotes everytime they visit the page. This comes from a local API
I developed. In order to use this from a git repo, users will need to
uses "npm start". This will also be in the README.